resolve some code format question

+++ common/types.go
@@ -147,7 +147,7 @@ func (h UnprefixedHash) MarshalText() ([]byte, error) {
 // return the string data that has been added to the num
 func IncreaseHexByNum(indexKeyHash []byte, num int64) string {
 	x := new(big.Int).SetBytes(indexKeyHash)
-	y := big.NewInt(int64(num))
+	y := big.NewInt(num)
 	x.Add(x, y)
 	return hex.EncodeToString(x.Bytes())
 }
